table: refactor render helper function to separate file (#242)
